+++ geo/src/algorithm/coords_iter.rs
@@ -30,6 +30,8 @@ pub trait CoordsIter<'a, T: CoordinateType> {
     /// assert_eq!(None, iter.next());
     /// ```
     fn coords_iter(&'a self) -> Self::Iter;
+
+    fn coords_count(&'a self) -> usize;
 }
 
 // ┌──────────────────────────┐
@@ -41,6 +43,11 @@ impl<'a, T: CoordinateType> CoordsIter<'a, T> for Point<T> {
 
     fn coords_iter(&'a self) -> Self::Iter {
         iter::once(self.0)
+    }
+
+    /// Return the number of coordinates in the `Point`.
+    fn coords_count(&'a self) -> usize {
+        1
     }
 }
 
@@ -54,6 +61,11 @@ impl<'a, T: CoordinateType> CoordsIter<'a, T> for Line<T> {
     fn coords_iter(&'a self) -> Self::Iter {
         iter::once(self.start).chain(iter::once(self.end))
     }
+
+    /// Return the number of coordinates in the `Line`.
+    fn coords_count(&'a self) -> usize {
+        2
+    }
 }
 
 // ┌───────────────────────────────┐
@@ -65,6 +77,11 @@ impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for LineString<T> {
 
     fn coords_iter(&'a self) -> Self::Iter {
         self.0.iter().copied()
+    }
+
+    /// Return the number of coordinates in the `LineString`.
+    fn coords_count(&'a self) -> usize {
+        self.0.len()
     }
 }
 
@@ -84,6 +101,12 @@ impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for Polygon<T> {
             .coords_iter()
             .chain(MapCoordsIter(self.interiors().iter(), marker::PhantomData).flatten())
     }
+
+    /// Return the number of coordinates in the `Polygon`.
+    fn coords_count(&'a self) -> usize {
+        self.exterior().num_coords() +
+            self.interiors().iter().map(|i| i.num_coords()).sum::<usize>()
+    }
 }
 
 // ┌───────────────────────────────┐
@@ -95,6 +118,11 @@ impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for MultiPoint<T> {
 
     fn coords_iter(&'a self) -> Self::Iter {
         MapCoordsIter(self.0.iter(), marker::PhantomData).flatten()
+    }
+
+    /// Return the number of coordinates in the `MultiPoint`.
+    fn coords_count(&'a self) -> usize {
+        self.0.len()
     }
 }
 
@@ -108,6 +136,11 @@ impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for MultiLineString<T> {
     fn coords_iter(&'a self) -> Self::Iter {
         MapCoordsIter(self.0.iter(), marker::PhantomData).flatten()
     }
+
+    /// Return the number of coordinates in the `MultiLineString`.
+    fn coords_count(&'a self) -> usize {
+        self.0.iter().map(|line_string| line_string.num_coords()).sum()
+    }
 }
 
 // ┌─────────────────────────────────┐
@@ -120,6 +153,11 @@ impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for MultiPolygon<T> {
     fn coords_iter(&'a self) -> Self::Iter {
         MapCoordsIter(self.0.iter(), marker::PhantomData).flatten()
     }
+
+    /// Return the number of coordinates in the `MultiPolygon`.
+    fn coords_count(&'a self) -> usize {
+        self.0.iter().map(|polygon| polygon.coords_count()).sum()
+    }
 }
 
 // ┌───────────────────────────────────────┐
@@ -131,6 +169,11 @@ impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for GeometryCollection<T> {
 
     fn coords_iter(&'a self) -> Self::Iter {
         Box::new(self.0.iter().flat_map(|geometry| geometry.coords_iter()))
+    }
+
+    /// Return the number of coordinates in the `GeometryCollection`.
+    fn coords_count(&'a self) -> usize {
+        self.0.iter().map(|geometry| geometry.coords_count()).sum()
     }
 }
 
@@ -164,6 +207,14 @@ impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for Rect<T> {
             y: self.min().y,
         }))
     }
+
+    /// Return the number of coordinates in the `Rect`.
+    ///
+    /// Note: Although a `Rect` is represented by two coordinates, it is
+    /// spatially represented by four, so this method returns `4`.
+    fn coords_count(&'a self) -> usize {
+        4
+    }
 }
 
 // ┌─────────────────────────────┐
@@ -177,6 +228,11 @@ impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for Triangle<T> {
         iter::once(self.0)
             .chain(iter::once(self.1))
             .chain(iter::once(self.2))
+    }
+
+    /// Return the number of coordinates in the `Triangle`.
+    fn coords_count(&'a self) -> usize {
+        3
     }
 }
 
@@ -201,6 +257,22 @@ impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for Geometry<T> {
             }
             Geometry::Rect(g) => GeometryCoordsIter::Rect(g.coords_iter()),
             Geometry::Triangle(g) => GeometryCoordsIter::Triangle(g.coords_iter()),
+        }
+    }
+
+    /// Return the number of coordinates in the `Geometry`.
+    fn coords_count(&'a self) -> usize {
+        match self {
+            Geometry::Point(g) => g.coords_count(),
+            Geometry::Line(g) => g.coords_count(),
+            Geometry::LineString(g) => g.coords_count(),
+            Geometry::Polygon(g) => g.coords_count(),
+            Geometry::MultiPoint(g) => g.coords_count(),
+            Geometry::MultiLineString(g) => g.coords_count(),
+            Geometry::MultiPolygon(g) => g.coords_count(),
+            Geometry::GeometryCollection(g) => g.coords_count(),
+            Geometry::Rect(g) => g.coords_count(),
+            Geometry::Triangle(g) => g.coords_count(),
         }
     }
 }